+++ src/Bicep.Local.Extension.UnitTests/Bicep.Local.Extension.UnitTests.csproj
@@ -23,4 +23,8 @@
     <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="HostTests\" />
+  </ItemGroup>
+
 </Project>
