+++ src/Bicep.Local.Extension.UnitTests/TypesTests/TypeProviderTests.cs
@@ -6,11 +6,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Bicep.Local.Extension.Host.Attributes;
-using Bicep.Local.Extension.Host.TypeDefinitionBuilder;
+using Bicep.Local.Extension.Types;
+using Bicep.Local.Extension.Types.Attributes;
 using FluentAssertions;
 
-namespace Bicep.Local.Extension.UnitTests.HostTests.TypeDefinitionBuilderTests.Types_A
+namespace Bicep.Local.Extension.UnitTests.TypesTests
+
 {
     [BicepType(IsActive = true)]
     public class ActiveResource { }
@@ -24,7 +25,7 @@ namespace Bicep.Local.Extension.UnitTests.HostTests.TypeDefinitionBuilderTests.T
     internal class InternalActiveResource { }
 }
 
-namespace Bicep.Local.Extension.UnitTests.HostTests.TypeDefinitionBuilderTests.Types_B
+namespace Bicep.Local.Extension.UnitTests.TypesTests.TypeDefinitionBuilderTests.Types_B
 {
     [BicepType(IsActive = true)]
     public class ActiveResource { }
@@ -38,7 +39,7 @@ namespace Bicep.Local.Extension.UnitTests.HostTests.TypeDefinitionBuilderTests.T
     internal class InternalActiveResource { }
 }
 
-namespace Bicep.Local.Extension.UnitTests.HostTests.TypeDefinitionBuilderTests
+namespace Bicep.Local.Extension.UnitTests.TypesTests.TypeDefinitionBuilderTests
 {
     [BicepType(IsActive = true)]
     public class ActiveResource { }
@@ -83,19 +84,19 @@ namespace Bicep.Local.Extension.UnitTests.HostTests.TypeDefinitionBuilderTests
             // users will have to implement their own ITypeProvider to handle
             // such scenarios. The default TypeProvider will only accept the first
             // occurance of the class name
-            types.Should().NotContain(typeof(Types_A.ActiveResource));
+            types.Should().NotContain(typeof(TypesTests.TypeDefinitionBuilderTests.ActiveResource));
             types.Should().NotContain(typeof(Types_B.ActiveResource));
 
             types.Should().NotContain(typeof(InternalActiveResource));
             types.Should().NotContain(typeof(PrivateNestedActiveResource));
 
             types.Should().NotContain(typeof(InactiveResource));
-            types.Should().NotContain(typeof(Types_A.InactiveResource));
+            types.Should().NotContain(typeof(TypesTests.TypeDefinitionBuilderTests.InactiveResource));
             types.Should().NotContain(typeof(Types_B.InactiveResource));
             types.Should().NotContain(typeof(NestedInactiveResource));
             types.Should().NotContain(typeof(NoAttributeResource));
             types.Should().NotContain(typeof(NestedNoAttributeResource));  
-            types.Should().NotContain(typeof(Types_A.ActiveResource));
+            types.Should().NotContain(typeof(TypesTests.TypeDefinitionBuilderTests.ActiveResource));
             types.Should().NotContain(typeof(Types_B.ActiveResource));          
         }
     }
