+++ src/Bicep.Local.Extension/Bicep.Local.Extension.csproj
@@ -28,4 +28,8 @@
   <ItemGroup>
     <PackageReference Update="Nerdbank.GitVersioning" Version="3.7.112" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Host\" />
+  </ItemGroup>
 </Project>
