+++ src/Bicep.Local.Extension/Types/Attributes/BicepTypeAttribute.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Bicep.Local.Extension.Host.Attributes;
+namespace Bicep.Local.Extension.Types.Attributes;
 
 [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
 public class BicepTypeAttribute : Attribute
