+++ src/Bicep.Local.Extension/Types/ITypeProvider.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Bicep.Local.Extension.Host.TypeDefinitionBuilder;
+namespace Bicep.Local.Extension.Types;
 public interface ITypeProvider
 {
     Type[] GetResourceTypes();
